Move Discord logic behind BustOutput protocol
With the end goal of decoupling the implementation of BustController
from the concrete Discord interface, we move most "User Interface"
operations and their precise output formats behind a BustOutput
interface, choosing instead to communicate through raw data.

The protocol isn't completely stateless as it keeps track of the pinned
message from the last song info display. We might change this in the
future, not sure.