Decouple discord audio player from BustController (#237)
Hide the concrete details of Discord audio playing behind an AudioPlayer
protocol. This also delegates the responsibility of connecting to and
disconnecting from the correct channel to the /bust caller. In the
future we might clean this up slightly as I don't love that the bust
caller has this responsibility; maybe make BustOutput a greater discord
interaction layer and ask it for the active channel? But for now, this
is still a vast improvement.